complete jwt

+++ src/main/java/com/techway/service/impl/ReplyCommentService.java
@@ -1,32 +1,51 @@
 package com.techway.service.impl;
 
 import java.util.List;
+import java.util.stream.Collectors;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.techway.exception.ResourceNotFoundException;
 import com.techway.model.dto.ReplyCommentDto;
 import com.techway.model.entity.Comment;
 import com.techway.model.entity.ReplyComment;
 import com.techway.model.entity.User;
 import com.techway.repository.CommentRepository;
 import com.techway.repository.ReplyCommentRepository;
+import com.techway.repository.RoleRepository;
 import com.techway.repository.UserRepository;
 import com.techway.service.IReplyCommentService;
 
 @Service
 public class ReplyCommentService implements IReplyCommentService{
+//	@Autowired
+//	HttpServletRequest requestServlet;
 	@Autowired
 	ReplyCommentRepository replyCommentRepository;
 	@Autowired
 	UserRepository userRepository;
 	@Autowired
 	CommentRepository commentRepository;
+	@Autowired
+	RoleRepository roleRepository;
+	
+	
+//	String email = requestServlet.getUserPrincipal().getName();
 	
 	@Override
-	public List<ReplyComment> findAllByUserId(Long userId) {
-		return replyCommentRepository.findAllByUserId(userId);
+	public List<ReplyCommentDto> findAllByUserId(Long userId) {
+		return replyCommentRepository.findAllByUserId(userId).stream().map(
+				reply -> mapEntityToDto(reply)
+				).collect(Collectors.toList());
+	}
+	
+	@Override
+	public List<ReplyCommentDto> findAllByCommentId(Long commentId) {
+		return replyCommentRepository.findAllByUserId(commentId).stream().map(
+				reply -> mapEntityToDto(reply)
+				).collect(Collectors.toList());
 	}
 	
 	@Override
@@ -37,21 +56,23 @@ public class ReplyCommentService implements IReplyCommentService{
 	
 
 	@Override
-	public boolean delete(long id) {
-		try {
-			ReplyComment reply = replyCommentRepository.findById(id).get();
+	public boolean delete(long replyId) {	
+//		User user = userRepository.findByEmail().get();
+		User user = new User();
+		ReplyComment reply = replyCommentRepository.findById(replyId).get();
+		if((user.getId() == reply.getUser().getId()) || 
+				user.getRoles().contains(roleRepository.findByName("ROLE_DIRE").get())) {
 			replyCommentRepository.delete(reply);
 			return true;
-		} catch (Exception e) {
-			e.printStackTrace();
+		}else
 			return false;
-		}
 	}
 	
 	private ReplyComment mapDtoToEntity(ReplyCommentDto request) {
 		ReplyComment response = new ReplyComment();
 		Comment comment = commentRepository.findById(request.getCommentId()).get();
-		User user = userRepository.findById(request.getUserId()).get();		
+//		User user = userRepository.findByEmail(email).get();
+		User user = new User();
 		response.setContent(request.getContent());
 		response.setComment(comment);
 		response.setUser(user);
@@ -66,5 +87,5 @@ public class ReplyCommentService implements IReplyCommentService{
 		response.setUserId(request.getUser().getId());
 		return response;
 	}
-
+	
 }
